perf: Home 페이지 접근 시 서버에서 데이터 prefetch하여 초기 렌더링 속도 개선
